+++ src-tauri/src/database/model/value.rs
@@ -1,11 +1,9 @@
 #![allow(dead_code)]
-use secrecy::SecretString;
+use super::*;
 use serde::de::{self, Deserializer, MapAccess, SeqAccess, Visitor};
-use serde::{Deserialize, Serialize};
 use std::fmt;
 use std::ops::Not;
 use totp_rs::{Rfc6238, TOTP};
-use zeroize::{Zeroize, ZeroizeOnDrop};
 
 /// Number value
 #[derive(Debug, PartialEq, Default, Zeroize, ZeroizeOnDrop, Serialize)]
