+++ src-tauri/src/lib.rs
@@ -1,11 +1,30 @@
+mod command;
 mod database;
-mod frontend;
-mod utils;
+mod totp;
+mod window;
 
-use frontend::*;
+use command::authentication::*;
+use command::cloud::*;
+use command::database::*;
+use command::password::*;
+use command::totp::*;
+use command::validation::*;
+use command::window::*;
+use command::*;
 use tauri::async_runtime::block_on;
-use tauri::Manager;
+use tauri::{AppHandle, Manager, Window};
 use tauri_plugin_window_state::{AppHandleExt, StateFlags};
+use totp::TOTPManager;
+
+/// Shows a critical error message and restarts the application.
+fn critical_error(message: &str, app_handle: AppHandle, window: Window) {
+    tauri::api::dialog::blocking::message(
+        Some(&window),
+        "Critical Error",
+        format!("{}\nApplication will now restart", message),
+    );
+    app_handle.restart();
+}
 
 /// Payload for single instance plugin.
 #[derive(Clone, serde::Serialize)]
@@ -28,7 +47,7 @@ pub fn run() -> anyhow::Result<()> {
             app.emit_all("single-instance", Payload { args: argv, cwd })
                 .unwrap_or_default();
         }))
-        .manage(TOTPManager::default())
+        .manage(TOTPManager::new(50))
         .invoke_handler(tauri::generate_handler![
             initialize_window,
             login,
@@ -42,9 +61,10 @@ pub fn run() -> anyhow::Result<()> {
             get_content_value,
             delete_content,
             get_totp_code,
-            valid,
+            validate,
             password_strength,
             check_password,
+            check_password_from_database,
             copy_value_to_clipboard,
             card_type,
             generate_password,
@@ -58,9 +78,7 @@ pub fn run() -> anyhow::Result<()> {
 
     block_on(initialize_window(app.app_handle()))?;
 
-    app.run(|_app_handle, _event| {
-        // Can react to events
-    });
+    app.run(|_app_handle, _event| { /* Can react to events */ });
 
     Ok(())
 }
